Fix some issues
There are some updates needed for the guard

Add a flashlight.

Add a guard model from Mixamo along with animations.

The guard behavior needs to be fixed  it should only follow the player when the guard sees the player not all the time.

When the guard catches the player, show a Game Ove screen with a try Again option.